Server/requirements.txt
Django==1.10.1
gunicorn==19.6.0
psycopg2==2.6.2
Pillow==2.9.0
djangorestframework==3.4.7
